+++ crates/runner-nordic/src/board/crypto.rs
@@ -22,10 +22,12 @@ pub enum Impl {}
 impl crypto::Api for Impl {
     #[cfg(feature = "aes128-ccm")]
     type Aes128Ccm = ccm::Impl;
+    #[cfg(feature = "software-crypto-aes256-cbc")]
+    type Aes256Cbc = crypto::SoftwareAes256Cbc;
     #[cfg(feature = "software-crypto-aes256-gcm")]
     type Aes256Gcm = crypto::SoftwareAes256Gcm;
     #[cfg(feature = "software-crypto-hmac-sha256")]
-    type HmacSha256 = crypto::SoftwareHmacSha256<Self>;
+    type HmacSha256 = crypto::SoftwareHmacSha256<Impl>;
     #[cfg(feature = "software-crypto-p256")]
     type P256 = crypto::SoftwareP256<Impl>;
     #[cfg(feature = "software-crypto-sha256")]